add blur image
