+++ app/lib/portfolio/etc.ts
@@ -10,36 +10,55 @@ import rehypeSlug from "rehype-slug";
 import LinkIcon from "@/app/components/markdown/LinkIcon";
 import LinkIconGrid from "@/app/components/markdown/LinkIconGrid";
 import {getAllMdFiles, parseMarkdown} from "@/app/lib/etc";
+import {getPlaiceholder} from "plaiceholder";
 
 const cwd = process.cwd();
 const projectsDirectory = join(cwd, '/_projects');
 
 export const getAllProjects = async () => {
 	const mdFiles = getAllMdFiles(projectsDirectory);
-	
-	const projects: ProjectType[] = [];
+
+	let projects: ProjectType[] = [];
 	for (const file of mdFiles) {
 		projects.push(await parseMarkdown(join(projectsDirectory, file)) as ProjectType);
 	}
-	
+
+	projects = await Promise.all(
+		projects.map(async (item, _) => {
+			let blurDataURL = "";
+			try {
+				if (item.thumbnail) {
+					const file = await fs.readFileSync(join("public", item.thumbnail));
+					const { base64 } = await getPlaiceholder(file);
+
+					blurDataURL = base64;
+				}
+			} catch (e) {
+				console.error(e);
+			}
+
+			return { ...item, blurDataURL };
+		})
+	)
+
 	projects.sort((a, b) => {
 		if (!a.displayPriority) return 1;
 		if (!b.displayPriority) return -1;
-		
+
 		return parseInt(a.displayPriority) - parseInt(b.displayPriority);
 	});
-	
+
 	return projects;
 }
 
 export const getProject = async (slug: string) => {
 	const parsed = decodeURIComponent(slug);
-	
+
 	const markdownWithMetadata = fs.readFileSync(
 		join(projectsDirectory, `${parsed}.mdx`),
 		'utf-8'
 	);
-	
+
 	const { content, frontmatter } = await compileMDX({
 		source: markdownWithMetadata,
 		options: {
@@ -55,7 +74,7 @@ export const getProject = async (slug: string) => {
 			LinkIcon
 		},
 	});
-	
+
 	return {
 		...frontmatter,
 		content,
