+++ package.json
@@ -13,6 +13,7 @@
     "@mdx-js/loader": "^2.3.0",
     "@mdx-js/react": "^2.3.0",
     "@next/mdx": "^13.4.5",
+    "@plaiceholder/next": "^3.0.0",
     "@types/mdx": "^2.0.5",
     "@types/node": "^20.2.5",
     "@types/react": "18.2.6",
@@ -24,6 +25,7 @@
     "gray-matter": "^4.0.3",
     "next": "^13.4.8-canary.2",
     "next-mdx-remote": "^4.4.1",
+    "plaiceholder": "^3.0.0",
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-device-detect": "^2.2.3",
